% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add.tree.grouping}
\alias{add.tree.grouping}
\title{Group a tree by a column from a data.frame.}
\usage{
add.tree.grouping(
  tree,
  metadata,
  tip.label.column = "tip.label",
  grouping.column
)
}
\arguments{
\item{tree}{the tree to group}

\item{metadata}{a data.frame with columns for tip labels and group}

\item{tip.label.column}{the name of the column in `metadata` with the tip
labels}

\item{grouping.column}{the name of the column in `metadata` with the grouping
variables}
}
\value{
the tree with the added grouping
}
\description{
The metadata data.frame provided should have a column with the tip labels in
the tree. The tip labels will be grouped by the values in grouping.column.
The tree group will have the same name as the grouping column.
}
\examples{
tree <- ape::read.tree(text = "(((a:1, b:1):1, (c:1, d:1):2):1, e:3);")
metadata <- data.frame(
  tip.label = tree$tip.label,
  Group = c("Group 1", "Group 1", "Group 2", "Group 2", "Group 3")
)
tree <- add.tree.grouping(tree, metadata,
  tip.label.column = "tip.label", grouping.column = "Group"
)
ggtree::ggtree(tree) +
  ggtree::geom_tree() +
  ggtree::geom_tiplab(ggtree::aes(col = Group))

}
