% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{biostrings.aln.to.list}
\alias{biostrings.aln.to.list}
\title{Convert a Biostrings MSA to a character list}
\usage{
biostrings.aln.to.list(aln)
}
\arguments{
\item{aln}{the Biostrings MSA}
}
\value{
a character list
}
\description{
Convert a Biostrings MSA to a character list format that can be exported as FASTA
}
