% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{calc.kaks}
\alias{calc.kaks}
\title{Given a nucleitide alignment file, calculate the KaKs and return in tidy format}
\usage{
calc.kaks(nt.aln.file)
}
\arguments{
\item{nt.aln.file}{the alignment file}
}
\value{
the KaKs ratio in long format
}
\description{
Given a nucleitide alignment file, calculate the KaKs and return in tidy format
}
