% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check.binary.exists}
\alias{check.binary.exists}
\title{If the given binary is not on the PATH, stop the script}
\usage{
check.binary.exists(binary)
}
\arguments{
\item{binary}{the binary to test}
}
\value{
nothing; halts the script if the binary is not found
}
\description{
If the given binary is not on the PATH, stop the script
}
\examples{
check.binary.exists("macse")
}
