% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert.to.gapped.coordinate}
\alias{convert.to.gapped.coordinate}
\title{Translate ungapped coordinates back to gapped coordinates}
\usage{
convert.to.gapped.coordinate(site.no.gap, gapped.seq)
}
\arguments{
\item{site.no.gap}{the integer site in an ungapped sequence to convert}

\item{gapped.seq}{the sequence with gaps from an alignment}
}
\value{
the cooordinate in the gapped sequence
}
\description{
Translate ungapped coordinates back to gapped coordinates
}
