% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.on.PATH}
\alias{is.on.PATH}
\title{Test if the given binary name is on the PATH}
\usage{
is.on.PATH(binary)
}
\arguments{
\item{binary}{the binary to search for}
}
\value{
true if the binary is on the PATH, false otherwise
}
\description{
Returns true if the binary is on the PATH, false otherwise
}
\examples{
is.on.PATH("macse")
}
