% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{matrix.aln.to.list}
\alias{matrix.aln.to.list}
\title{Convert a matrix MSA to a character list format}
\usage{
matrix.aln.to.list(aln)
}
\arguments{
\item{aln}{the MSA as a matrix}
}
\value{
a character list
}
\description{
Convert a matrix MSA to a character list format
}
