% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.kaks}
\alias{plot.kaks}
\title{Calculate and plot KaKs}
\usage{
\method{plot}{kaks}(nt.aln.file, species.order, kaks.limits = c(0, 1))
}
\arguments{
\item{nt.aln.file}{the nucleotide alignment}

\item{species.order}{a vector with the plotting order for species in the file}

\item{kaks.limits}{min and max for the plot KaKs scale}
}
\description{
Given an alignment and order of species, calculate KaKs and make a pairwise plot
}
