% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.tree}
\alias{plot.tree}
\title{Plot a tree as read by ggtree or ape. The tree may have bootstrap values in
the node labels}
\usage{
plot.tree(tree, tiplab.font.size = 2, ...)
}
\arguments{
\item{tree}{the tree}

\item{tiplab.font.size}{the size of tip text in ggtree}

\item{...}{other parameters to geom_tiplab}
}
\description{
Plot a tree as read by ggtree or ape. The tree may have bootstrap values in
the node labels
}
