% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printMultipleAlignment}
\alias{printMultipleAlignment}
\title{Convert FASTA format to clustal style format.}
\usage{
printMultipleAlignment(
  alignment,
  names = NULL,
  names.length = NA,
  chunksize = 60
)
}
\arguments{
\item{alignment}{seqinr alignment format}

\item{names}{optional character vector of names (if null, alignment names are used)}

\item{names.length}{override the default width of the names column (if na, default is used)}

\item{chunksize}{number of letters per row}
}
\value{
the formatted text
}
\description{
Prints and returns as a text object
}
