% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.clustalo.dual.profile}
\alias{run.clustalo.dual.profile}
\title{Create a alignment from two existing alignments}
\usage{
run.clustalo.dual.profile(aln.file.1, aln.file.2, out.file)
}
\arguments{
\item{aln.file.1}{the first alignment file}

\item{aln.file.2}{the second alignment file (or single sequence FASTA)}

\item{out.file}{the combined output alignment file}
}
\description{
Create a alignment of two existing alignments (or an alignment and a single
sequence FASTA) This preserves the structure of the existing alignments.
Expects 'clustalo' on the PATH.
}
