% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.divvier}
\alias{run.divvier}
\title{Run divvier from within a conda environment}
\usage{
run.divvier(aln.file, ...)
}
\arguments{
\item{aln.file}{the alignment to divvy}

\item{...}{other options to divvier}
}
\value{
the divvied alignment file path
}
\description{
Run divvier from within a conda environment (named divvier).Creates a divvied
alignment.
}
