% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.iqtree}
\alias{run.iqtree}
\title{Run IQTREE}
\usage{
run.iqtree(aln.file, ...)
}
\arguments{
\item{aln.file}{the alignment}

\item{...}{other arguments to IQTREE}
}
\value{
the path to the treefile
}
\description{
Expects 'iqtree3' on the PATH. By default, runs with 6 threads
}
