% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.macse.align.sequences}
\alias{run.macse.align.sequences}
\title{Run MACSE alingment on an input FASTA file}
\usage{
run.macse.align.sequences(fa.file, out.dir = "aln", ...)
}
\arguments{
\item{fa.file}{the FASTA file to align}

\item{out.dir}{the directory to save the alignment files}

\item{...}{other arguments to macse}
}
\value{
a list with the alignment file names for AA and NT
}
\description{
Outputs NT and AA alignment files. Expects 'macse' on the PATH
}
