% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.macse.export.alignment}
\alias{run.macse.export.alignment}
\title{Remove stop codons from NT alignment.}
\usage{
run.macse.export.alignment(nt.aln.file, out.dir = "aln", ...)
}
\arguments{
\item{nt.aln.file}{aligned FASTA file containing stop codons and or frameshifts.}

\item{out.dir}{the directory to save the alignment files}

\item{...}{other arguments to macse}
}
\value{
a list with the alignment file names for AA, NT and the stats file
}
\description{
Using macse exportAlignment. Expects 'macse' on the PATH.
}
