% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.R
\name{run.macse}
\alias{run.macse}
\title{Generic wrapper for MACSE}
\usage{
run.macse(...)
}
\arguments{
\item{...}{all arguments to macse}
}
\description{
Run arbitrary MACSE commands. The subprogram and all options must be
specified.
}
