% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.macse.trim.non.homologous.fragments}
\alias{run.macse.trim.non.homologous.fragments}
\title{Trim non-homologous sequences from an unaligned nucleotide sequence}
\usage{
run.macse.trim.non.homologous.fragments(fa.file, out.dir = "aln", ...)
}
\arguments{
\item{fa.file}{the FASTA file to trim}

\item{out.dir}{the directory to save the trimmed files. Defaults to 'aln'.}

\item{...}{other arguments to macse}
}
\value{
a list with the trimmed file names for NT and the output stats
}
\description{
Using MACSE `trimNonHomologousFragments`. Uses default parameters of
min_homology_to_keep_seq 0.2, min_trim_in 40 and min_trim_ext 20. Expects
'macse' on the PATH
}
\examples{
# trim.files <- run.macse.trim.non.homologous.fragments(cds.file)
}
