% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.mafft}
\alias{run.mafft}
\title{Run MAFFT on an input FASTA}
\usage{
run.mafft(fa.file, out.dir = "aln", ...)
}
\arguments{
\item{fa.file}{the file of sequences to align}

\item{out.dir}{the directory to save the alignment file}

\item{...}{other arguments to mafft}
}
\value{
the alignment file name
}
\description{
Outputs data/seq.fa to aln/seq.aln
}
