% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.muscle}
\alias{run.muscle}
\title{Run muscle}
\usage{
run.muscle(fa.file, out.dir = "aln", ...)
}
\arguments{
\item{fa.file}{the FASTA file to align}

\item{out.dir}{the directory to output the alignment. Defaults to 'aln'}

\item{...}{other arguments to muscle}
}
\value{
the path to the alignment file
}
\description{
Expects 'muscle' on the PATH.
}
