% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.ncbi.datasets}
\alias{run.ncbi.datasets}
\title{Download ortholgoues for the given gene symbol using NCBI datasets}
\usage{
run.ncbi.datasets(
  symbol,
  ortholog = "all",
  include = "cds",
  overwrite = FALSE,
  filename,
  ...
)
}
\arguments{
\item{symbol}{the gene symbol to search}

\item{ortholog}{which orthologues to
retrieve}

\item{include}{the sequence types to return. Defaults to cds}

\item{overwrite}{should we overwrite existing downloads. If false, and the outputs  exist, the function will return the output zip file name. Defaults to false.}

\item{filename}{the path for the downloaded zip file}

\item{...}{other arguments to datasets}
}
\value{
a vector of FASTA file paths
}
\description{
Download ortholgoues for the given gene symbol using NCBI datasets. Extracts
the downloaded file, and moves sequence files to folder containing the zip
file. Expects 'datasets' on the PATH
}
