% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save.double.width}
\alias{save.double.width}
\title{Save a double column plot (170mm) at 600dpi}
\usage{
save.double.width(filename, plot = ggplot2::last_plot(), height = 170)
}
\arguments{
\item{filename}{the file to save to}

\item{plot}{the plot to save. Defaults to the last generated plot.}

\item{height}{the height in mm. Defaults to 170mm (square plot).}
}
\description{
Save a double column plot (170mm) at 600dpi
}
