% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save.plot}
\alias{save.plot}
\title{Save a plot to file}
\usage{
save.plot(filename, plot = ggplot2::last_plot(), width, height)
}
\arguments{
\item{filename}{the file to save to}

\item{plot}{the plot to save. Defaults to the last generated plot.}

\item{width}{the width in mm}

\item{height}{the height in mm}
}
\description{
Saves png and svg versions of the plot
}
