% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.R
\name{stop.if.missing}
\alias{stop.if.missing}
\title{If the given binary is not on the PATH, stop the script}
\usage{
stop.if.missing(binary)
}
\arguments{
\item{binary}{the binary to test}
}
\value{
nothing; halts the script if the binary is not found
}
\description{
If the given binary is not on the PATH, stop the script
}
\examples{
stop.if.missing("macse")
}
