% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subset.sequence}
\alias{subset.sequence}
\title{Extract sequence from a Biostrings alignment.}
\usage{
\method{subset}{sequence}(aln, sequence.name, start, end)
}
\arguments{
\item{aln}{msa in Biostrings format}

\item{sequence.name}{the name of the sequence from the alignment to extract}

\item{start, end}{coordinates in the gapped alignment}
}
\value{
the sequence from the alignment
}
\description{
Given a Biostrings alignment, extract the sequence string at the given coordinates
}
