% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update.tip.labels}
\alias{update.tip.labels}
\title{Change the tip labels of a tree to a column from a data.frame.}
\usage{
update.tip.labels(tree,
  metadata,
  current.tip.label.column = "current.tip.label",
  new.tip.label.column = "new.tip.label")
}
\arguments{
\item{tree}{the tree to update}

\item{metadata}{a data.frame with columns for current and new tip labels}

\item{current.tip.label.column}{the name of the column in `metadata` with the
current tip labels}

\item{new.tip.label.column}{the name of the column in `metadata` with the new
tip labels}
}
\value{
the tree with the tip labels updated to the new values
}
\description{
The metadata data.frame provided should have columns with the current tip
labels and the new tip labels in the tree. The tip labels will be updated,
and the original labels added to a new vector in the tree
}
\examples{
tree <- ape::read.tree(text = "(((a:1, b:1):1, (c:1, d:1):2):1, e:3);")
metadata <- data.frame(
  CurrentLabel = tree$tip.label,
  NewLabel = c("1", "2", "3", "4", "5")
)
tree <- update.tip.labels(tree, metadata,
  current.tip.label.column = "CurrentLabel",
  new.tip.label.column = "NewLabel"
)
plot(tree)
}
