% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create.xlsx}
\alias{create.xlsx}
\title{Create an Excel file containing a data frame}
\usage{
create.xlsx(data, file.name, cols.to.fixed.size.font = NULL)
}
\arguments{
\item{data}{the data to export}

\item{file.name}{the name of the file}

\item{cols.to.fixed.size.font}{integer vector of columns to set in monospace font}
}
\description{
Create an Excel file containing a data frame
}
