% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reroot.tree}
\alias{reroot.tree}
\title{Reroot the given tree to the given tip labels}
\usage{
reroot.tree(tree, node.labels, position = 0.01)
}
\arguments{
\item{tree}{the tree to reroot (e.g. ape)}

\item{node.labels}{vector of tips to root on}

\item{position}{the position to place the root on the selected branch}
}
\value{
the rerooted tree
}
\description{
If the tip labels contains more than one node, their MRCA will be used as the root.
}
