% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run.hyphy.meme}
\alias{run.hyphy.meme}
\title{Run HyPhy MEME in a Unix environment}
\usage{
run.hyphy.meme(nex.file, tree.file)
}
\arguments{
\item{nex.file}{the nexus file with the alignment}

\item{tree.file}{the tree for the alignment}
}
\value{
the output file path
}
\description{
Tests the given test and background branches in the tree file.
Assumes a conda environment named 'hyphy' is available
}
