% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaries.R
\name{run.iqtree3}
\alias{run.iqtree3}
\title{Run IQTREE3}
\usage{
run.iqtree3(aln.file, overwrite = FALSE, ...)
}
\arguments{
\item{aln.file}{the alignment}

\item{overwrite}{overwrite existing output files only if TRUE}

\item{...}{other arguments to IQTREE}
}
\value{
the path to the treefile
}
\description{
Expects 'iqtree3' on the PATH. By default, runs with 6 threads
}
